Element with maximum tabIndex cannot be returned by nextElementWithGreaterTabIndex() https://bugs.webkit.org/show_bug.cgi?id=155215
Reviewed by Ryosuke Niwa.

Source/WebCore:

Element with maximum tabIndex cannot be returned by nextElementWithGreaterTabIndex()
due to a bug in r197726. This patch fixes the issue by only comparing
candidate.tabIndex to winningTabIndex if winner is non-null.

Test: fast/events/max-tabindex-focus.html

* page/FocusController.cpp:
(WebCore::nextElementWithGreaterTabIndex):

LayoutTests:

Add test to make sure that an Element with a tabIndex equal to
2147483647 (maximum tabIndex) can be focused.

* fast/events/max-tabindex-focus-expected.txt: Added.
* fast/events/max-tabindex-focus.html: Added.


Identifier: 173325@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@197835 268f45cc-cd09-0410-ab3c-d52691b4dbfc