Set the WebContent process's main thread QoS to USER-INTERACTIVE https://bugs.webkit.org/show_bug.cgi?id=155595 <rdar://problem/22534965>
Reviewed by Antti Koivisto.

Source/WebKit2:

Increase the WebContent process main thread's QoS to USER-INTERACTIVE
instead of USER-INITIATED as it is drawing UI. However, use a relative
priority of -1 so that its priority is lower than the one of the
scrolling thread.

* WebProcess/WebProcess.cpp:
(WebKit::WebProcess::initializeWebProcess):

Source/WTF:

Add a relativePriority parameter to setCurrentThreadIsUser*() so that
we can do more fine-grained prioritization of threads that have the
same QoS.

* wtf/Threading.cpp:
(WTF::setCurrentThreadIsUserInteractive):
(WTF::setCurrentThreadIsUserInitiated):
(WTF::createThread): Deleted.
* wtf/Threading.h:

Identifier: 173712@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@198350 268f45cc-cd09-0410-ab3c-d52691b4dbfc