Find-in-page indicator in Mail viewer is the wrong scale and cut off https://bugs.webkit.org/show_bug.cgi?id=155605 <rdar://problem/23948165>
Reviewed by Simon Fraser.

Source/WebKit2:

* UIProcess/mac/WKTextFinderClient.mm:
(-[WKTextFinderClient didGetImageForMatchResult:]):
Initialize the NSImage with the correct size, instead of inferring the
size from the bitmap, so that we don't lose information about device pixel ratio.

Tools:

* TestWebKitAPI/Tests/WebKit2Cocoa/FindInPage.mm:
(TEST):
Add a test that ensures that the find result image is correctly @2x.
The NSImage size should be in points.


Identifier: 173724@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@198363 268f45cc-cd09-0410-ab3c-d52691b4dbfc