Data URL DecodeTask may get deleted outside main thread https://bugs.webkit.org/show_bug.cgi?id=155584 rdar://problem/24492104
Reviewed by Darin Adler.

This is unsafe as it owns strings and other types that are only safe to delete in the main thread.

There is a race between deref in dispatch() and deref in timerFired(). If the timer fires before dispatch()
exits the implicit deref will trigger deletion of DecodingResultDispatcher in the dispatching thread.

(WebCore::DataURLDecoder::DecodingResultDispatcher::timerFired):

    Fix by clearing m_decodeTask when the timer fires.



Identifier: 173744@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@198387 268f45cc-cd09-0410-ab3c-d52691b4dbfc